GCP関係の処理をGCP.goに移動
